requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
Flask-SQLAlchemy==3.0.2
SQLAlchemy==2.2.5
psycopg2-binary==2.9.10
python-dotenv==1.0.0
bcrypt==4.0.1
